Update to have all challenges use files
This also includes refactoring to asking the user for input instead of
having them provide it via command line args. It dynamically figures out
how many challenges there are and prompts the user based on that number.
It's worth noting that if we change how the challenge folders are named,
we'll have to revisit this logic and change it accordingly.